Broken

+++ include/ft_parse.h
@@ -1,9 +1,11 @@
 #ifndef FT_PARSE_H
 #define FT_PARSE_H
 
+#include <stdbool.h>
+#include <stdint.h>
+
 #include "../include/ft_arena.h"
 #include "../include/ft_array.h"
-#include <stdbool.h>
 
 typedef struct {
     bool list;
@@ -13,6 +15,6 @@ typedef struct {
     bool time;
 } t_args;
 
-t_array *parse_args(Arena *arena, int argc, char **argv, t_args *args);
+t_array *parse_args(Arena *arena, uint64_t argc, char **argv, t_args *args);
 
 #endif // !FT_PARSE_H
